requirements.txt
alembic==0.8.8
aniso8601==1.1.0
backports-abc==0.4

click==6.6
decorator==4.0.10
entrypoints==0.2.2
factory-boy==2.7.0
fake-factory==0.7.2
Flask==0.11.1

flask-marshmallow==0.7.0
Flask-Migrate==2.0.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
Mako==1.0.4
MarkupSafe==0.23
marshmallow==2.10.3
marshmallow-sqlalchemy==0.12.0
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.3
numpy==1.11.2
pandas==0.19.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.31
Pygments==2.1.3
pytest==3.0.3
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.7
pyzmq==16.0.0
qtconsole==4.2.1
requests==2.11.1
requests-mock==1.1.0
shortuuid==0.4.3
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.2
terminado==0.6
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.11
widgetsnbextension==1.2.6
xmltodict==0.10.2
